First Version of IVAQ QGS
Monitoring Panel of IVAQ Payload added:
1) Front-end made in FlightDisplayViewIVAQPayload. To do: Change layout to Rectancle with a Grid/Column layout inside with Rows for the different parameters and columns.
Bring Color definitions to the frontend

2) Back End in src/IVAQ/IVAQPayloadManager.h/.c .
To Do: Named_Value_Float variables need to be parsed too. We need to see if we need a chart plotting the signal or if we can use the Mavlink Inspector. We also need to detect that the board is not responding / powered off.
D